Win32: improved MinGW/MinGW-w64 GCC checks.
Previously, __GNUC__ was checked, which is now replaced with __MINGW32__
checks.  The difference is that __MINGW32__ is defined when using MinGW
(or MinGW-w64) header files regardless of the compiler being used.  And,
more importantly, it is not defined when Clang is being used (which
pretends to be GCC by default) with Windows SDK header files.

With this change, it is now possible to compile nginx with native Clang
on Windows.  This current requires --with-ld-opt="-lkernel32 -luser32"
though, since native Clang on Windows uses MSVC linker, which does
not link kernel32.lib and user32.lib automatically.