Support for Clang with "-fgnuc-version=0".
Previously, the code assumed that Clang pretends to be GCC and defines
the __GNUC__ macro, which might not be the case if the "-fgnuc-version=0"
flag is used.  Fix is to explicitly check for __clang__ as well.

In practice, with this change ngx_cpuinfo() is now available on Linux
with "-fgnuc-version=0".