SSL: added additional verify context check for OpenSSL.
When using TLSv1.3, OpenSSL 1.1.1e+ allows session resumption with names
other than initially negotiated, and provides no documented way to prevent
it or even detect.  This makes it possible to resume the session with
a certificate verified in the context of a different server block, similarly
to 5095:4fbef397c753.

There is no such problem in LibreSSL since it doesn't support session
resumption with TLSv1.3 (and does not allow session resumption with
different names with TLSv1.2 and below).  Similarly, there is no such
problem with BoringSSL.  While BoringSSL allows sessions to be resumed
with different names, it does so only after checking session id context
set by the servername callback.

With this change, SSL_get_servername() (which, for TLSv1.3, returns name
as sent by the client in the current handshake) is checked against
SSL_SESSION_get0_hostname(), and thus prevents incorrect use of certificates
verified in different contexts.

Note that SSL_SESSION_get0_hostname() is documented to return NULL with
TLSv1.3, but in practice it returns the originally negotiated name at least
up to OpenSSL 3.4.1, and the documentation is expected to be fixed.