Support for Multipath TCP on Linux.
With this change, Multipath TCP on Linux can be activated with
"listen ... multipath".  The "listen ... multipath" option is supported
in http, stream, and mail modules.

Note that on Linux to activate Multipath TCP one should create a socket
with the IPPROTO_MPTCP protocol explicitly specified, and it is not possible
to change an existing socket.  To make transition possible with minimal
impact on client connections, if the multipath option is changed in the
configuration, SO_REUSEPORT is set on the old socket, and the new socket
is opened with IPPROTO_MPTCP.

Note that this creates a race window, and connection requests which are
assigned to the old socket will be lost.  In particular, this might affect
binary upgrade when the WINCH signal is used to preserve the old master
process.

Requires Linux kernel 5.6 or newer.  Note though that some of the socket
options might not be supported with Multipath TCP or only supported in
new kernels.  Most notably, TCP_NODELAY is only supported with kernel
version 5.17 or newer.

Based on patches by Maxime Dourov and Anthony Doeraene.