Upstream: fixed passwords usage for certificates with variables.
SSL certificate passwords are stored separately from the SSL context
created for SSL proxying, yet modified when the context is created
if certificates with variables are used (to ensure passwords will be
available at run time).  Optimizations introduced in 8053:9d98d524bd02
(1.23.1) did not take this into account, and might end up using at run
time passwords which weren't preserved to be usable at run time, such as
in the following configuration:

server {
    proxy_ssl_certificate $crt;
    proxy_ssl_certificate_key $key;
    proxy_ssl_password_file foo;

    location /1/ {
        proxy_pass https://u;
    }

    location /2/ {
        proxy_pass https://u;
    }
}

Fix is to preserve passwords if needed when using an inherited SSL
context.