Compatibility with -Wunterminated-string-initialization in GCC 15.
Builds with recently released GCC 15 currently fail due to
-Wunterminated-string-initialization (introduced in -Wextra), which warns
about attempts to initialize a character array with a string literal
if the trailing NUL character doesn't fit (perfectly valid in C, but
might be unintentional).

Fix is to use arrays for initialization instead of string literals, or
switch to using character arrays of unknown size where appropriate.