Stream: fixed timeout usage for proxy_protocol with SSL proxying.
Connection establishment with a proxied server includes SSL handshake,
and the relevant timeout is set with the proxy_connect_timeout directive.
Since proxy_protocol sending happens before SSL handshake, it should
use proxy_connect_timeout as well.  Further, the timeout should not be
rearmed if it's already set.

Additionally, read handler should be set as long as we are waiting
for events, since it can be triggered.