Mp4: of.directio_off flag to avoid unneeded syscalls.
When set, ngx_open_cached_file() won't try to enable directio on the file,
allowing the caller to do it itself, but will set is_directio flag in
ngx_open_file_info_t and in open file cache.  The fact that directio needs
to be enabled by the caller is signalled in the of.is_directio_off output
flag.

This approach makes it possible to avoid unneeded syscalls when directio
is enabled, yet the caller needs it disabled for some initial file
processing, such as in the mp4 module.