Changed limit_rate algorithm to leaky bucket.
The bucket size (burst size) is set to the size expected to be sent
in 1 second (that is, the rate configured), so the first operation will
be able to send corresponding number of bytes, matching the previous
behaviour.  To ensure that inaccurate timers won't affect rate limiting
accuracy, delays are set to trigger next sending when the bucket is
half-empty.

The limit_rate_after directive makes the bucket size larger, allowing
for larger traffic bursts (it won't affect delays though).  This is mostly
equivalent to the previous behaviour in simple cases, like downloading
a static file by a fast client, but expected to work better in complex
cases, such as when actual response traffic might stop for a while.

Similar changes are made to proxy_limit_rate (and friends) in the http
module, as well as proxy_upload_rate and proxy_download_rate in the stream
module.

Immediate benefits include more accurate limiting, notably at rates which
resulted in 1 or 2 millisecond delays after sending (and very inaccurate
limiting) with the old algorithm.  Further, there will be less unneeded
delays, notably no delays at all in most cases as long as the client is
not using allowed bandwidth.

Note that due to the new algorithm there are minor changes in the limiting,
mostly visible with very small limits: typically 1.5x of the limit will
be sent in the first second (one immediately, and half after 0.5 seconds
delay), and additional data will be sent once per 0.5 seconds (instead
of once per second previously).  Some tests needs to be adapted for this.

Note that r->limit_last is not initialized, and will be set to an actual
value only after first sending.  As a result, "ms" can be large during
initial sending, and to ensure that there will be no signed integer
overflows, calculations of "excess" are performed in (uint64_t) (and the
result is type casted into (off_t) to clarify that the code intentionally
relies on implementation-defined behaviour, similarly to how we handle
calculations of "ms").