Added the "send_min_rate" configuration directive.
Specifies the minimum required data transmission rate to the client.
If the average transmission rate is below this value for send_timeout
period, the connection is closed.

The default value is 0, which implies that any data transfer results
in a new timeout (that is, the existing behaviour).

Algorithm used is as follows: the send_timeout timer is only rearmed
if average sending rate after the timer was last rearmed is above the
specified minimum.

Note that this directive might interfere with limit_rate if it is used:
minimum rates below ((limit_rate / 2 + limit_rate_after) / send_timeout)
might not work, due to delay timers used by limit_rate on c->write.
To avoid interference, using limit_rate_after at least as large as
(send_min_rate * send_timeout) is recommended.