Mail: limit_conn module.
The limit_conn module limits connections after auth_http authentication,
so it can use the user name updated by the authentication server, and
headers returned by the authentication server.  Mostly usable to limit
number of connections from a particular authenticated client.

Configuration is similar to the one in HTTP, but since there are no
variables in mail, limit_conn_zone can use the following restricted
set of keys:

- $remote_addr, client address;
- $remote_user, user name supplied during authentication;
- $auth_http_*, auth_http response header field.

For example:

    limit_conn_zone $auth_http_limit_key zone=one:1m;
    limit_conn one 10;

If the limit is hit, the following errors are returned for IMAP, POP3,
and SMTP (and the connection is closed):

    <tag> NO [UNAVAILABLE] Too many connections
    -ERR [SYS/TEMP] Too many connections
    454 4.7.0 Too many connections

For IMAP, the "[UNAVAILABLE]" response code (RFC 5530) is used.

For POP3, the "[SYS/TEMP]" extended response code (RFC 2449, RFC 3206)
is used, and the default capabilities list is updated with RESP-CODES
to ensure that extended response codes can be properly recognized by
clients.

At least Apple Mail (both on macOS and iOS) is able to recognize the
response codes being used, and won't show password prompt if the limit
is hit.  Still, many clients won't be able to.  In particular, Thunderbird
is known to show a generic error dialog with options to try again or
enter a new password, and Outlook always asks for a new password.
As such, it is not recommended to configure limits which are low enough
to be hit by well-behaving clients.