Fixed left-shifting of signed integers into the sign bit.
If a bitwise left shift operation is used on a signed operand, and the
resulting value cannot be represented in the result type, the behaviour
is undefined.

In all cases in question the result is converted to the appropriate
unsigned type, and this doesn't cause any issues with known compilers.
Still, avoiding undefined behaviour is generally a good idea, only needs
a cast, and we already do this in multiple other places.  See also
6627:ad736705a744 and 6626:b3682580c1bd.

Prodded by UndefinedBehaviorSanitizer, "-fsanitize=shift".