Proxy: reworked HTTP/0.9 handling.
Previous behaviour was to downgrade the client connection to HTTP/0.9
and return the upstream response as is.  However, this approach doesn't
really work with HTTP/2, and will break HTTP/3 as currently implemented
(as it relies on r->http_version).  Further, this approach is potentially
unsafe: if a status line is not recognized as HTTP/1.x by nginx, but
accepted by the client, the response can be recognized as multiple
responses.

Now, instead of downgrading client connection to HTTP/0.9, we generate
a 200 response with the response body received via HTTP/0.9, much like
the memcached module does.

Additionally, now logging is done consistently at the "debug" level,
and special handling for cache is removed.  This resolves various
issues with cache enabled, such as missing status in $upstream_status
and no logging at all.