Proxy: HTTP/0.9 responses now disabled by default.
Compatibility with HTTP/0.9 is no longer essential nowadays.  Further,
it often causes confusion when a malformed HTTP/1.x response is interpreted
as a HTTP/0.9 response, or when proxying to a non-HTTP server appears to
work and return something.

As such, compatibility with HTTP/0.9 responses in the proxy module is
now disabled by default.  The "proxy_allow_http09" directive makes it
possible to re-enable it if needed in the particular setup.