Upstream: simplified ngx_http_upstream_process_header().
This restores the simple loop structure which was somewhat screwed up
in 9237:41db21d1ca7c, and also removes some commented out artifacts of
the past.

Note that checking c->read->ready before the first c->recv() is not
required, since ngx_http_upstream_process_header() is an event handler
and it cannot be called multiple times in a loop.