Upstream: unexpected connection upgrades now rejected.
Unless the client explicitly requested to change the application protocol,
the "101 Switching Protocols" response is now considered to be an upstream
server error and not forwarded to the client.

Similarly, other 1xx responses are also rejected for now.  This will be
changed in subsequent patches.

This ensures that such responses won't affect the connection with the client.