Upstream: unexpected 1xx interim responses now ignored.
Unexpected 1xx interim informational responses received before the final
response are now parsed and ignored.  This is required for HTTP/1.1
and above, and expected to provide better interoperability.

Notable exception is 101 (Switching Protocols), which is rejected unless
requested by the client, as it is expected to be followed by other
protocol data which won't be able to parse anyway.  Similarly, invalid
responses with status codes below 100 are rejected as well.

Amount of 1xx responses which can be received from an upstream server
and ignored is generally limited by the buffer size.  For gRPC, since
buffer can be reset between reading HEADERS frames, total number of
1xx responses is additionally limited.  The limit is set to 10, which
happens to match the limit used by Apache.