Updated request line parsing to allow IPv6 zone identifiers.
RFC 6874 defines syntax to allow IPv6 zone identifiers in URIs, by
providing an additional option in IP-literal:

  IP-literal = "[" ( IPv6address / IPv6addrz / IPvFuture  ) "]"
  ZoneID = 1*( unreserved / pct-encoded )
  IPv6addrz = IPv6address "%25" ZoneID

The "%" character wasn't previously allowed in IP literals, and therefore
attempts to use such addresses in request line resulted in 400 (Bad Request)
(but was accepted in the Host header field, which uses more relaxed parsing).

With this change, "%" is now allowed in IP literals.