Updated request line parsing to allow uncommon chars in host.
Previously, only ALPHA, DIGIT, ".", and "-" were allowed in the host
component of the request line (if it's not an IP literal).  On the other
hand, RFC 3986 allows the following:

  reg-name    = *( unreserved / pct-encoded / sub-delims )
  unreserved  = ALPHA / DIGIT / "-" / "." / "_" / "~"
  sub-delims  = "!" / "$" / "&" / "'" / "(" / ")"
              / "*" / "+" / "," / ";" / "="
  pct-encoded = "%" HEXDIG HEXDIG

Notably, the "_" character is used in practice but was not allowed in the
request line.  At the same time, this and other characters do actually work
in practice, as they are accepted in the Host header field, which uses more
relaxed parsing.

With this change, all characters which are valid in the host name per
RFC 3986 are also allowed in the request line.