Improved ngx_http_validate_host() to do better syntax validation.
With this change, syntax validation in ngx_http_validate_host() is mostly
identical to the one used for the request line parsing, which now follows
RFC 3986.  In particular, as a result of this change non-ASCII characters
and invalid port numbers are no longer allowed in the Host header.

This also fixes "." not being stripped if port contained ".", as such
requests are now rejected by syntax checks.

Prodded by Anton Mironenko,
https://github.com/freenginx/nginx/issues/11