Geo: fixed ngx_file_info() error handling.
When ngx_file_info() on the text include file failed when loading a binary
include, the code failed to restore binary file name in "name".  While
it is generally not needed after an error, an additional error might
theoretically happen when closing the binary include, and wrong "name"
will result in incorrect file name in the error message.

The fix is to restore "name", much like it is done in the normal code path.

Prodded by Coverity (CID 1306890, 1643342).