GeoIP: support for GeoIP2 databases in the MMDB format.
With this change, the ngx_http_geoip_module can be compiled with the
GeoIP library (legacy), or the libmaxminddb library, or both.  Usage
of specific libraries can be disabled during compilation with new configure
options, "--without-geoip" and "--without-libmaxminddb".

Existing directives (geoip_country, geoip_org, geoip_city) support both
file formats, and existing variables are mapped to the appropriate data
paths in the MaxMind GeoIP2 databases, for example:

    geoip_country /path/to/GeoLite2-Country.mmdb;

Notable exceptions: the $geoip_country_code3, $geoip_city_country_code3,
and $geoip_area_code variables are no longer available if GeoIP2 databases
are used, because the underlying fields are not available in the new
databases.

Additionally, the geoip_set directive is introduced.  It can be used to
access arbitrary data paths within an MMDB database, for example:

    geoip_set $country /path/to/GeoLite2-Country.mmdb country.names.en;
    geoip_set $asn /path/to/GeoLite2-ASN.mmdb autonomous_system_number;

The last parameter is the data path to look up in the database, similarly
to the one accepted by the mmdblookup utility (but with dots between
path components).