GeoIP: rearranged checks to avoid dead code.
The "ngx_http_geoip_mmdb_file() == NGX_OK" checks are not executed when
both the legacy GeoIP library and libmaxminddb available, and Coverity
complains that the code is dead.  Fix is to rearrange to code so the
checks are not compiled at all if both libraries are available.

Prodded by Coverity.